# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for a fitness application",
        "title": "MFit API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/workouts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetch all workout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workouts"
                ],
                "summary": "Fetch all workout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/store.PresentableWorkout"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Creates a workout",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workouts"
                ],
                "summary": "Creates a workout",
                "parameters": [
                    {
                        "description": "Workout payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.CreateWorkoutPayload"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/store.Workout"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {}
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        },
        "/workouts/{workoutId}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Fetches a workout by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "workouts"
                ],
                "summary": "Fetches a workout",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Workout ID",
                        "name": "workoutId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/store.PresentableWorkout"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {}
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {}
                    }
                }
            }
        }
    },
    "definitions": {
...
        "main.CreateWorkoutPayload": {
            "type": "object",
            "required": [
                "bodypart_id",
                "difficulty",
                "equipment_id",
                "name",
                "primary_target",
                "secondary_targets"
            ],
            "properties": {
                "bodypart_id": {
                    "type": "integer"
                },
                "calories_burned": {
                    "type": "integer"
                },
                "difficulty": {
                    "type": "string",
                    "enum": [
                        "beginner",
                        "intermediate",
                        "advanced"
                    ]
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "equipment_id": {
                    "type": "integer"
                },
                "gif_url": {
                    "type": "string"
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string",
                    "maxLength": 40
                },
                "primary_target": {
                    "type": "integer"
                },
                "secondary_targets": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "store.PresentableWorkout": {
            "type": "object",
            "properties": {
                "body_part": {
                    "type": "string"
                },
                "calories_burned": {
                    "type": "integer"
                },
                "difficulty": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "equipment": {
                    "type": "string"
                },
                "gif_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                },
                "primary_target": {
                    "type": "string"
                },
                "secondary_targets": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "store.Workout": {
            "type": "object",
            "properties": {
                "bodypart_id": {
                    "type": "integer"
                },
                "calories_burned": {
                    "type": "integer"
                },
                "difficulty": {
                    "type": "string"
                },
                "duration_minutes": {
                    "type": "integer"
                },
                "equipment_id": {
                    "type": "integer"
                },
                "gif_url": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "instructions": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        }
    }
}